requirements.txt
grpcio
grpcio-tools
flask
requests
matplotlib
